refactored the bulk actions form step 1 into a table for re-use
